chore: define tasks schema
