+++ apps/app/docs/README.md
@@ -3,3 +3,9 @@
 Guidelines and design references for the user-facing application.
 
 Include screenshots of wireframes and note any design tokens used from `@repo/design-system`.
+
+## Routes
+
+- `/` - main authenticated workspace page (previously `/code`).
+- `/settings` - workspace settings page.
+- `/c/[id]` - view a task's interaction history and messages.
