--- apps/app/src/app/(authenticated)/code/layout.tsx
@@ -1,13 +1,0 @@
-import { Metadata } from 'next';
-
-export const metadata: Metadata = {
-  title: 'arbor code',
-};
-
-export default function CodeLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return <>{children}</>;
-}
