+++ apps/app/src/app/(authenticated)/layout.tsx
@@ -1,8 +1,13 @@
+import { Metadata } from "next";
 import { env } from '@/env';
 import { auth, currentUser } from '@repo/auth/server';
 import { SidebarProvider } from '@repo/design-system/components/ui/sidebar';
 import { NotificationsProvider } from '@repo/notifications/components/provider';
 import { secure } from '@repo/security';
+export const metadata: Metadata = {
+  title: "arbor code"
+};
+
 import type { ReactNode } from 'react';
 import { PostHogIdentifier } from './components/posthog-identifier';
 import { GlobalSidebar } from './components/sidebar';
