+++ packages/database/docs/README.md
@@ -13,4 +13,10 @@ Prisma client configured for Neon serverless. Exports `database` instance and ge
 ```ts
 import { database } from '@repo/database'
 ```
-Access models via `database.user.findMany()` and similar methods.
+Access models via `database.task.findMany()` and similar methods.
+
+## Models
+
+- `Task` – stores the submitted URL, optional prompt, status, and creation timestamps.
+- `Message` – represents tool calls or text output associated with a task.
+- `TaskStatus` and `MessageType` enums define available states.
