+++ packages/database/prisma/schema.prisma
@@ -1,5 +1,4 @@
-// This is your Prisma schema file,
-// learn more about it in the docs: https://pris.ly/d/prisma-schema
+// Prisma schema for Webs
 
 generator client {
   provider        = "prisma-client-js"
@@ -13,9 +12,35 @@ datasource db {
   relationMode = "prisma"
 }
 
-// This is a stub model.
-// Delete it and add your own Prisma models.
-model Page {
-  id   Int    @id @default(autoincrement())
-  name String
+model Task {
+  id        String       @id @default(uuid())
+  url       String
+  prompt    String?
+  status    TaskStatus   @default(PENDING)
+  createdAt DateTime     @default(now())
+  updatedAt DateTime     @updatedAt
+  messages  Message[]
+}
+
+model Message {
+  id        String       @id @default(uuid())
+  taskId    String?
+  task      Task?        @relation(fields: [taskId], references: [id])
+  type      MessageType  @default(TEXT)
+  content   String
+  createdAt DateTime     @default(now())
+}
+
+enum TaskStatus {
+  PENDING
+  RUNNING
+  COMPLETE
+  FAILED
+}
+
+enum MessageType {
+  TEXT
+  TOOL
+  SYSTEM
+  AI
 }
